moving scripts
